requirements.txt
aiogram
openai
python-decouple
redis
DateTime
requests
